28-CHANGE: File restructure & some rewrites
