+++ .github/workflows/ci.yml
@@ -9,7 +9,7 @@ on:
 
 jobs:
   build:
-    name: Build and Test
+    name: Build
     timeout-minutes: 10
     runs-on: ubuntu-latest
     # To use Remote Caching, uncomment the next lines and follow the steps below.
@@ -28,6 +28,7 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: 16
+          cache: "yarn"
 
       - name: Install dependencies
         run: yarn
@@ -35,7 +36,40 @@ jobs:
       - name: Run Build
         run: yarn build
 
-      - name: Run Tests
+  start:
+    name: Start up Meally
+    runs-on: ubuntu-latest
+    needs: build
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 2
+
+      - name: Setup Node.js environment
+        uses: actions/setup-node@v3
+        with:
+          node-version: 16
+          cache: "yarn"
+
+      - name: Start Meally
+        run: yarn start
+
+  test&lint:
+    name: Build
+    timeout-minutes: 10
+    runs-on: ubuntu-latest
+    needs: build
+
+    steps:
+      - name: Setup Node.js environment
+        uses: actions/setup-node@v3
+        with:
+          node-version: 16
+          cache: "yarn"
+
+      - name: Test
         run: yarn test
 
-      
+      - name: Lint
+        run: yarn lint
