+++ GETTING_STARTED.md
@@ -23,7 +23,7 @@ start just the one you want
 
 ```bash
     npm run dev --filter *app name*
-    // e.g npm run dev --filter meally.com.au
+    # e.g npm run dev --filter meally.com.au
 ```
 
 ## Contributing
@@ -36,7 +36,8 @@ If you are new to the project and want to help out but don't know where to start
 
 ## file structure
 
-The repo uses a custom Nextjs 13 file structure with [Turborepo](https://turbo.build/repo). This means that the repo is a monorepo with multiple apps and packages. The apps are the actual websites and the packages are the shared code between the apps. The apps and packages are all in the same repo but are separate from each other. This means that the apps and packages can be developed and tested independently of each other. This also means that the apps and packages can be published independently of each other. This is very useful for the packages as they can be published to npm and used in other projects.
+The repo uses a custom Nextjs 13 file structure with [Turborepo](https://turbo.build/repo). The project has been set up as a monorepo with multiple apps and packages. 
+All apps live under the [apps directory](./apps/) and packages that are shared between apps are under [packages](./packages/).
 
 ## Workspaces
 
@@ -47,8 +48,7 @@ If you aren't familiar with Workspaces they are pretty much ways of separating w
 Through this project there is mostly a consistent code style so it is easy to follow and understand no matter where you are in the project. When working on the project it would be greatly appreciated if you could follow this style.
 
 ### Passing props
-
-When passing props to a component i like to have this sort of structure:
+Passing props
 
 ```jsx
     interface componentProps {
