+++ SECURITY.md
@@ -1,13 +1,13 @@
 # Security Policy
-
-Apart of the security we strive to keep our user safe while on our platform however from time to time without us knowing about it a vulnerability shows up. If you have found any vulnerabilities than please report it Read below for more info, however if this bug isn't fixed in **90 days** than feel free to open it to the world.
+Apart of the security we strive to keep our user safe while on our platform however from time to time without us knowing about it a vulnerability shows up. If you have found any vulnerabilities than please report it Read below for more info. 
 
 ## Reporting a Vulnerability
+First and for most **DON'T** report it under Bugs
 
 When reporting a vulnerability please [Email Me](jacob35422@gmail.com) straight away so i can get to fixing it ASAP, when doing so please explain the following:
 
-1. How it occurred,
-2. What is compromised,
-3. Steps to replicate the issue
+1. How it occurred
+2. Steps to replicate the issue
+3. What is \ could be compromised
 
 Thankyou for taking your time to read this and help keep Meally safe
