--- apps/docs/.storybook/main.js
@@ -1,25 +1,0 @@
-module.exports = {
-  stories: ["../stories/**/*.stories.mdx", "../stories/**/*.stories.tsx"],
-  addons: ["@storybook/addon-links", "@storybook/addon-essentials"],
-  framework: "@storybook/react",
-  core: {
-    builder: "@storybook/builder-vite",
-  },
-  // async viteFinal(config, { configType }) {
-  //   // customize the Vite config here
-  //   return {
-  //     ...config,
-  //     resolve: {
-  //       alias: [
-  //         {
-  //           find: "@acme/core",
-  //           replacement: path.resolve(
-  //             __dirname,
-  //             "../../../packages/acme-core/"
-  //           ),
-  //         },
-  //       ],
-  //     },
-  //   };
-  // },
-};
