--- apps/docs/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "docs",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "devDependencies": {
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-links": "^6.5.16",
-    "@storybook/react": "^6.5.16",
-    "@storybook/testing-library": "^0.0.13",
-    "eslint-config-custom": "*",
-    "serve": "^13.0.4",
-    "tsconfig": "*",
-    "typescript": "4.7.4"
-  },
-  "dependencies": {
-    "@storybook/builder-vite": "^7.0.0",
-    "libs": "*",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "ui": "*",
-    "vite": "^4.2.0"
-  },
-  "scripts": {
-    "dev": "start-storybook -p 6006",
-    "build": "build-storybook",
-    "clean": "rm -rf .turbo && rm -rf node_modules"
-  }
-}
