--- apps/meally.com.au/.dockerignore
@@ -1,39 +1,0 @@
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-/.VSCodeCounter
-/scripts
-
-# testing
-/coverage
-
-# next.js
-/.next/
-/out/
-.turbo
-
-# production
-/build
-
-# misc
-.DS_Store
-*.pem
-
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.pnpm-debug.log*
-
-# local env files
-.env*.local
-.env
-
-# vercel
-.vercel
-
-# typescript
-*.tsbuildinfo
-next-env.d.ts
-
