+++ apps/meally.com.au/src/common/styles/globals.scss
@@ -12,8 +12,6 @@ body {
   margin: 0;
   font-family: Roboto, Arial, -apple-system, Oxygen, Ubuntu, Cantarell,
     Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-  // width: 100%;
-  // height: 100%;
 }
 
 a {
