--- packages/styles/responsive_design.scss
@@ -1,39 +1,0 @@
-// Responsive type
-:root {
-  --step--4: clamp(0.24rem, calc(0.08rem + 0.83vw), 0.82rem);
-  --step--3: clamp(0.29rem, calc(0.08rem + 1.05vw), 1.02rem);
-  --step--2: clamp(0.35rem, calc(0.08rem + 1.33vw), 1.28rem);
-  --step--1: clamp(0.42rem, calc(0.08rem + 1.69vw), 1.60rem);
-  --step-0: clamp(0.50rem, calc(0.07rem + 2.14vw), 2.00rem);
-  --step-1: clamp(0.60rem, calc(0.06rem + 2.71vw), 2.50rem);
-  --step-2: clamp(0.72rem, calc(0.03rem + 3.44vw), 3.13rem);
-  --step-3: clamp(0.86rem, calc(-0.01rem + 4.35vw), 3.91rem);
-  --step-4: clamp(1.04rem, calc(-0.06rem + 5.49vw), 4.88rem);
-  --step-5: clamp(1.24rem, calc(-0.14rem + 6.94vw), 6.10rem);
-}
-
-// Responsive type space
-:root {
-  --space-3xs: clamp(0.13rem, calc(0.02rem + 0.54vw), 0.50rem);
-  --space-2xs: clamp(0.25rem, calc(0.04rem + 1.07vw), 1.00rem);
-  --space-xs: clamp(0.38rem, calc(0.05rem + 1.61vw), 1.50rem);
-  --space-s: clamp(0.50rem, calc(0.07rem + 2.14vw), 2.00rem);
-  --space-m: clamp(0.75rem, calc(0.11rem + 3.21vw), 3.00rem);
-  --space-l: clamp(1.00rem, calc(0.14rem + 4.29vw), 4.00rem);
-  --space-xl: clamp(1.50rem, calc(0.21rem + 6.43vw), 6.00rem);
-  --space-2xl: clamp(2.00rem, calc(0.29rem + 8.57vw), 8.00rem);
-  --space-3xl: clamp(3.00rem, calc(0.43rem + 12.86vw), 12.00rem);
-
-  /* One-up pairs */
-  --space-3xs-2xs: clamp(0.13rem, calc(-0.12rem + 1.25vw), 1.00rem);
-  --space-2xs-xs: clamp(0.25rem, calc(-0.11rem + 1.79vw), 1.50rem);
-  --space-xs-s: clamp(0.38rem, calc(-0.09rem + 2.32vw), 2.00rem);
-  --space-s-m: clamp(0.50rem, calc(-0.21rem + 3.57vw), 3.00rem);
-  --space-m-l: clamp(0.75rem, calc(-0.18rem + 4.64vw), 4.00rem);
-  --space-l-xl: clamp(1.00rem, calc(-0.43rem + 7.14vw), 6.00rem);
-  --space-xl-2xl: clamp(1.50rem, calc(-0.36rem + 9.29vw), 8.00rem);
-  --space-2xl-3xl: clamp(2.00rem, calc(-0.86rem + 14.29vw), 12.00rem);
-
-  /* Custom pairs */
-  --space-s-l: clamp(0.50rem, calc(-0.50rem + 5.00vw), 4.00rem);
-}
