+++ packages/ui/src/modules/AdBanner.tsx
@@ -36,7 +36,7 @@ function AdBanner() {
     );
   }
 
-  return <h1>Loading..</h1>;
+  return <h1>Loading...</h1>;
 }
 
 export { AdBanner };
