+++ packages/ui/src/modules/loader.tsx
@@ -1,6 +1,11 @@
 import React from "react";
 import "./loader.css";
 
+
+/**
+ * Spinning loader for when you are waiting for a response
+ * @returns Spinning loader
+ */
 const Loader = () => {
   return (
     <div className="pinwheel">
